Added vote reward system

+++ cogs/vote.py
@@ -28,30 +28,42 @@ class Vote(commands.Cog, name='Vote'):
             if message.channel.id == 902203209014190132 and message.guild.id == 828628284806922251:
                 usrid = int(message.content)
                 usr = self.client.get_user(usrid)
-                print(usr)
-                print(usr.id)
                 db = sqlite3.connect('cogs/main.sqlite')
                 cursor = db.cursor()
                 cursor.execute(
-                f"SELECT user FROM votes WHERE user = {message.content}")
+                f"SELECT user FROM votes WHERE user = '{message.content}'")
                 result = cursor.fetchone()
-                if result is None:
+                if result is not None:
                     return
                 sql = (
-                    "INSERT INTO votes(user) VALUES(?)")
-                val = (usr.id)
-                cursor.execute(sql, val)
+                    f"INSERT INTO votes(user) VALUES('{message.content}')")
+                cursor.execute(sql)
                 db.commit()
                 cursor.close()
                 db.close()
-                await usr.send(content="🍍 **| Thank you for voting for Pineapple Bot! You can now use `-claimvote` in a server to claim your vote reward!")
+                embed = discord.Embed(
+                    title="📩 Vote", description="**Thank you for voting!**\n\nUse `-vote claim` in a server to claim your reward!", color=0x0068d6)
+                embed.set_thumbnail(
+                    url=f"{str(self.client.user.avatar_url)}")
+                embed.set_footer(text=f"You can vote every 12 hours.")
+                await usr.send(embed=embed)
+                
         except Exception as e :
             print(f"{e.__class__.__name__}: {e}")
 
-    
-    @commands.command()
+    @commands.group(invoke_without_command=True)
     @commands.guild_only()
-    async def claimvote(self, ctx):
+    async def vote(self, ctx):
+        embed = discord.Embed(
+            title="📩 Vote", description="**You can vote for the bot on [Top.gg](https://top.gg/bot/463388759866474506).**", color=0x0068d6)
+        embed.set_thumbnail(
+            url=f"{str(self.client.user.avatar_url)}")
+        embed.set_footer(text=f"You can vote every 12 hours.")
+        await ctx.send(embed=embed)
+
+    @vote.command()
+    @commands.guild_only()
+    async def claim(self, ctx):
         def kform(num, round_to=2):
             if abs(num) < 1000:
                 return num
@@ -61,19 +73,18 @@ class Vote(commands.Cog, name='Vote'):
                     magnitude += 1
                     num = round(num / 1000.0, round_to)
                 return '{:.{}f}{}'.format(round(num, round_to), round_to, ['', 'K', 'M', 'G', 'T', 'P'][magnitude])
-        price = random.randint(900, 1500)
+        price = random.randint(1000, 10000)
         db = sqlite3.connect('cogs/main.sqlite')
         cursor = db.cursor()
         cursor.execute(
-        f"SELECT user FROM votes WHERE user = {ctx.author.id}")
+        f"SELECT user FROM votes WHERE user = '{str(ctx.author.id)}'")
         result = cursor.fetchone()
         if result is None:
             await ctx.send("<a:no:898507018527211540> **| You haven't voted yet! Use `-vote` to vote.**")
         else:
             sql = (
-                "DELETE FROM votes WHERE user = ?")
-            val = (ctx.author.id)
-            cursor.execute(sql, val)
+                f"DELETE FROM votes WHERE user = '{str(ctx.author.id)}'")
+            cursor.execute(sql)
             db.commit()
             cursor.execute(
                 f"SELECT user FROM economy WHERE user = {ctx.author.id}")
@@ -90,9 +101,9 @@ class Vote(commands.Cog, name='Vote'):
                 cursor.execute(sql, val)
                 db.commit()
             embed = discord.Embed(
-                title="🎁 Vote Reward", description=f"You just received <:silver:856609576459304961> **{price}** by voting", color=0x005ec2)
+                title="🎁 Vote Reward", description=f"You just received <:silver:856609576459304961> **{kform(price)}** for voting", color=0x005ec2)
             embed.set_thumbnail(url=ctx.author.avatar_url_as(size=256))
-            embed.set_footer(text=f"You can vote using -vote every 12 hours.")
+            embed.set_footer(text=f"{webs} | {ctx.author}")
             await ctx.send(embed=embed)
             cursor.close()
             db.close()
